backend/requirements.txt
asgiref==3.10.0
Django==5.2.7
dj-database-url==2.3.0
django-cors-headers==4.6.0
djangorestframework==3.16.1
gunicorn==23.0.0
psycopg[binary]==3.2.3
requests==2.32.3
sqlparse==0.5.3
whitenoise==6.8.2
openai==1.59.3